requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
openai==0.28.0
tiktoken==0.9.0
pypdf==5.3.0
aiofiles==24.1.0
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
PyPDF2==3.0.1
faiss-cpu==1.10.0
python-dotenv==1.0.1
-e . # this will look for setup.py in the current directory # modular approach
